Implemented Security layer
I implemented a Security system based on token authentication. I added the system to the existing create and delete files for asset.

+++ asset/delete.php
@@ -5,8 +5,20 @@ header("Access-Control-Allow-Methods: POST");
 header("Access-Control-Max-Age: 3600");
 header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
 include_once '../service/AssetService.php';
+include_once '../security/Security.php';
 
 // get product id
 $data = json_decode(file_get_contents("php://input"));
-
-AssetService::deleteOneById($data->id);
+if(!empty($data->token)) {
+    if (Security::authorizeUser(htmlspecialchars(strip_tags($data->token)))) {
+        AssetService::deleteOneById($data->id);
+    }
+    else {
+        http_response_code(503);
+        echo json_encode(array("message" => "User authentication failed. Invalid or expired token", "auth" => false));
+    }
+}
+else {
+    http_response_code(400);
+    echo json_encode(array("message" => "Authentication failed. Auth token missing.", "auth" => false));
+}
